backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
PyMySQL==1.1.0

# Django Channels for WebSocket
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# CORS handling
django-cors-headers==4.3.1

# Image processing
Pillow==10.2.0

# Filtering
django-filter==23.5

# Environment variables
python-decouple==3.8

# Async tasks (for notifications)
celery==5.3.4
django-celery-beat==2.5.0

# API documentation
drf-yasg==1.21.7

# Utilities
python-dateutil==2.8.2
